Play video feeds from all saved sources.
- Removed Exoplayer in favor of libVLC: Exoplayer refused to stream from
the PTZOptics cameras at church, whereas VLC proper would, so I'm hoping
libVLC also can stream from our cameras.
- We now instantiate a new VLC player + view for each camera that's
saved in our cameras repository. (Currently no way to edit this
repository outside of hacking in a `.addCamera()`, see lines 31-32 of
`MainActivity.kt` for examples with live-as-of-now test RTSP feeds).
Players are currently arranged in a `LinearLayout` horizontally with
equal widths, this will cease being optimal for n(cameras) > 3 or so,
but I don't care as I'm only planning to control 2 ;)